requirements.txt
faster-whisper>=1.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
pystray>=0.19.5
Pillow>=10.0.0
PyYAML>=6.0.1
torch>=2.0.0
torchaudio>=2.0.0
